requirements.txt
# ─── Core RAG Stack ────────────────────────────────────────────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3

# ─── Embeddings & Vector Store ─────────────────────────────────────────────────
sentence-transformers==3.0.1
faiss-cpu==1.8.0
huggingface-hub==0.24.6

# ─── LLM (free, runs locally via Hugging Face) ─────────────────────────────────
transformers==4.44.2
torch==2.4.1
accelerate==0.33.0

# ─── Document Loaders ──────────────────────────────────────────────────────────
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.15.12

# ─── Backend API ───────────────────────────────────────────────────────────────
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# ─── Frontend UI ───────────────────────────────────────────────────────────────
streamlit==1.38.0

# ─── Evaluation ────────────────────────────────────────────────────────────────
ragas==0.1.21
datasets==2.21.0

# ─── Experiment Tracking ───────────────────────────────────────────────────────
mlflow==2.15.1

# ─── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
